feat(webauthn): passkeys
Add support for passkeys, granular attachment modality, and granular authenticator selection. This also adds metadata verification elements utilizing the MDS3 to the project, including saving attestation statements, verification of attestation statements, etc. This also makes a significant change to the authentication level logic to purely use authentication method references to ensure the future-proof nature of the implementation. It should be noted that at the time of this commit Passkey authentication is considered a single factor and we will at a later stage add more customizable policies to allow configuring the requirements for a specific authentication level on a per rule basis using a custom 'policy' that administrators can define.

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>